+++ Sources/KeyboardKit/Appearance/KeyboardAppearance.swift
@@ -21,13 +21,23 @@ public protocol KeyboardAppearance {
     
     var keyboardBackgroundColor: Color { get }
     
-    func buttonBackgroundColor(for action: KeyboardAction, isPressed: Bool) -> Color
-    func buttonCornerRadius(for action: KeyboardAction) -> CGFloat
-    func buttonFont(for action: KeyboardAction) -> Font
-    func buttonForegroundColor(for action: KeyboardAction, isPressed: Bool) -> Color
-    func buttonImage(for action: KeyboardAction) -> Image?
-    func buttonShadowColor(for action: KeyboardAction) -> Color
-    func buttonText(for action: KeyboardAction) -> String?
-    
     func systemKeyboardButtonStyle(for action: KeyboardAction, isPressed: Bool) -> SystemKeyboardButtonStyle
+    
+    
+    // MARK: - Deprecated
+    
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonBackgroundColor(for action: KeyboardAction, isPressed: Bool) -> Color
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonCornerRadius(for action: KeyboardAction) -> CGFloat
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonFont(for action: KeyboardAction) -> Font
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonForegroundColor(for action: KeyboardAction, isPressed: Bool) -> Color
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonImage(for action: KeyboardAction) -> Image?
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonShadowColor(for action: KeyboardAction) -> Color
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    func buttonText(for action: KeyboardAction) -> String?
 }