+++ Sources/KeyboardKit/Appearance/StandardKeyboardAppearance.swift
@@ -26,36 +26,14 @@ open class StandardKeyboardAppearance: KeyboardAppearance {
     
     open var keyboardBackgroundColor: Color { .clear }
     
-    open func buttonBackgroundColor(for action: KeyboardAction, isPressed: Bool) -> Color {
-        action.standardButtonBackgroundColor(
-            for: context,
-               isPressed: isPressed)
-    }
-    
-    open func buttonCornerRadius(for action: KeyboardAction) -> CGFloat {
-        .standardKeyboardButtonCornerRadius(
-            for: context.device)
-    }
-    
     open func buttonFont(for action: KeyboardAction) -> Font {
         let rawFont = action.standardButtonFont(for: context)
         guard let weight = fontWeight(for: action) else { return rawFont }
         return rawFont.weight(weight)
     }
     
-    open func buttonForegroundColor(for action: KeyboardAction, isPressed: Bool) -> Color {
-        action.standardButtonForegroundColor(
-            for: context,
-               isPressed: isPressed)
-    }
-    
     open func buttonImage(for action: KeyboardAction) -> Image? {
         action.standardButtonImage(
-            for: context)
-    }
-    
-    open func buttonShadowColor(for action: KeyboardAction) -> Color {
-        action.standardButtonShadowColor(
             for: context)
     }
     
@@ -74,6 +52,34 @@ open class StandardKeyboardAppearance: KeyboardAppearance {
                 color: action.standardButtonShadowColor(for: context),
                 size: 1)
         )
+    }
+    
+    
+    
+    // MARK: - Deprecated
+    
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    open func buttonBackgroundColor(for action: KeyboardAction, isPressed: Bool) -> Color {
+        systemKeyboardButtonStyle(for: action, isPressed: isPressed)
+            .backgroundColor
+    }
+    
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    open func buttonCornerRadius(for action: KeyboardAction) -> CGFloat {
+        systemKeyboardButtonStyle(for: action, isPressed: false)
+            .cornerRadius
+    }
+    
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    open func buttonForegroundColor(for action: KeyboardAction, isPressed: Bool) -> Color {
+        systemKeyboardButtonStyle(for: action, isPressed: isPressed)
+            .foregroundColor
+    }
+    
+    @available(*, deprecated, message: "Use systemKeyboardButtonStyle instead")
+    open func buttonShadowColor(for action: KeyboardAction) -> Color {
+        systemKeyboardButtonStyle(for: action, isPressed: false)
+            .shadow.color
     }
 }